Integrate frontend with backend, implement SocketManager component, remove proto-plakoto logic
